Misc fixes to disk driver and real mode ints

+++ drivers/disk.h
@@ -1,9 +1,10 @@
 #ifndef __DISK_H__
 #define __DISK_H__
 
+#include <stddef.h>
 #include <stdint.h>
 
-int read_sector(int, void *, uint64_t, size_t);
-int read(int, void *, int, size_t);
+int read_sector(int, void *, uint64_t, uint64_t);
+int read(int, void *, uint64_t, uint64_t);
 
 #endif
